MAINT: Move code from _page.py to _text_extraction
